Make users injectable
