+++ app/src/main/java/ch/epfl/sweng/eventmanager/repository/data/User.java
@@ -1,19 +1,37 @@
 package ch.epfl.sweng.eventmanager.repository.data;
 
-public interface User {
+import com.google.firebase.auth.FirebaseUser;
+import com.google.firebase.database.*;
 
-    /**
-     * @return an unique ID representing the user
-     */
-    public String getUid();
+public final class User {
+    private final FirebaseDatabase db = FirebaseDatabase.getInstance();
+    private DatabaseReference userEntry;
 
-    /**
-     * @return a string representation of the user's common name
-     */
-    public String getDisplayName();
+    private String uid;
+    private String email;
+    private String displayName;
 
-    /**
-     * @return a string representation of the user's email
-     */
-    public String getEmail();
+    public User() {
+        // Used for dependency injection
+    }
+
+    public User(String uid, String email, String displayName) {
+        this.uid = uid;
+        this.email = email;
+        this.displayName = displayName;
+    }
+
+    public User(FirebaseUser firebaseUser) {
+        this.uid = firebaseUser.getUid();
+    }
+
+    public String getUid() { return uid; }
+
+    public String getDisplayName() {
+        return displayName;
+    }
+
+    public String getEmail() {
+        return email;
+    }
 }
