+++ app/src/main/java/ch/epfl/sweng/eventmanager/ui/user/UserManagerModule.java
@@ -13,5 +13,4 @@ public abstract class UserManagerModule {
 
     @ContributesAndroidInjector
     abstract DisplayAccountActivity contributeDisplayAccountActivityInjector();
-
 }
