+++ app/src/main/java/ch/epfl/sweng/eventmanager/users/DummyInMemorySession.java
@@ -5,7 +5,6 @@ import android.content.Intent;
 import com.google.android.gms.tasks.OnCompleteListener;
 import javax.inject.Singleton;
 
-import ch.epfl.sweng.eventmanager.repository.data.DummyUser;
 import ch.epfl.sweng.eventmanager.repository.data.User;
 import ch.epfl.sweng.eventmanager.ui.user.DisplayAccountActivity;
 
@@ -22,12 +21,12 @@ public class DummyInMemorySession implements InMemorySession{
     public static final String DUMMY_UID = "u0YmYQasWpNaNYZt4iXngV0aTxF3";
     public static final String DUMMY_DISPLAYNAME = "Lamb Da";
 
-    private DummyUser user;
+    private User user;
 
     @Override
     public void login(String email, String password, Activity context, OnCompleteListener callback) {
         if (email.equals(DUMMY_EMAIL) && password.equals(DUMMY_PASSWORD)) {
-            user = new DummyUser(DUMMY_UID,DUMMY_DISPLAYNAME, DUMMY_EMAIL);
+            user = new User(DUMMY_UID,DUMMY_DISPLAYNAME, DUMMY_EMAIL);
 
             // Switch to DisplayAccountActivity if successfully authenticated
             if (context != null) {
