+++ app/src/main/java/ch/epfl/sweng/eventmanager/users/InMemoryFirebaseSession.java
@@ -1,8 +1,8 @@
 package ch.epfl.sweng.eventmanager.users;
 
 import android.app.Activity;
-import ch.epfl.sweng.eventmanager.repository.data.FirebaseBackedUser;
 import ch.epfl.sweng.eventmanager.repository.data.User;
+
 import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.firebase.auth.FirebaseAuth;
 
@@ -11,7 +11,7 @@ import javax.inject.Singleton;
 @Singleton
 public class InMemoryFirebaseSession implements InMemorySession {
     private final FirebaseAuth mAuth;
-    private FirebaseBackedUser user;
+    private User user;
 
     public InMemoryFirebaseSession() {
         mAuth = FirebaseAuth.getInstance();
@@ -32,7 +32,7 @@ public class InMemoryFirebaseSession implements InMemorySession {
         if (mAuth.getCurrentUser() == null) {
             user = null;
         } else if (user == null || !user.getUid().equals(mAuth.getCurrentUser().getUid())) {
-            user = new FirebaseBackedUser(mAuth.getCurrentUser());
+            user = new User(mAuth.getCurrentUser());
         }
         return user;
     }
